requirements.txt
Flask==2.0.3
gunicorn==20.1.0
google-cloud-storage==1.42.3
pandas==1.3.4  
